Split test and bench into corresponding folders
 - Most of the code is not shared, so this simplifies updating/debugging only corresponding parts.
 - `cargo build --bench parser-bench` works and [profile.bench] applies instead of [profile.test].
 - Benchmarks can be still executed as integration tests.
