+++ tests/parser-test.rs
@@ -17,12 +17,11 @@ use estree::IntoScript;
 use glob::glob;
 use joker::track::Untrack;
 use serde_json::value::Value;
-use std::ffi::OsStr;
 use std::fs::{File, read_dir};
-use std::io::{Read, Write, stdout};
+use std::io::Read;
 use std::path::Path;
-use std::{thread, env};
-use test::{TestDesc, TestDescAndFn, TestName, TestFn, Bencher, TDynBenchFn, test_main};
+use std::env;
+use test::{TestDesc, TestDescAndFn, TestName, TestFn, test_main};
 use test::ShouldPanic::No;
 use unjson::{ExtractField, Unjson};
 
@@ -37,27 +36,6 @@ fn add_test<F: FnOnce() + Send + 'static>(tests: &mut Vec<TestDescAndFn>, name: 
     });
 }
 
-struct DynBenchFn<F> {
-    run: F
-}
-
-impl<F: Fn(&mut Bencher) + Send + 'static> TDynBenchFn for DynBenchFn<F> {
-    fn run(&self, harness: &mut Bencher) {
-        (self.run)(harness);
-    }
-}
-
-fn add_bench<F: Fn(&mut Bencher) + Send + 'static>(tests: &mut Vec<TestDescAndFn>, name: String, ignore: bool, f: F) {
-    tests.push(TestDescAndFn {
-        desc: TestDesc {
-            name: TestName::DynTestName(name),
-            ignore: ignore,
-            should_panic: No
-        },
-        testfn: TestFn::DynBenchFn(Box::new(DynBenchFn { run: f }))
-    });
-}
-
 fn as_ref_test(tests: &mut Vec<TestDescAndFn>) {
     add_test(tests, "reference test".to_string(), false, || {
         let mut ast = script("foobar = 17;").unwrap();
@@ -69,91 +47,6 @@ fn as_ref_test(tests: &mut Vec<TestDescAndFn>) {
             _ => { panic!("unexpected AST structure"); }
         }
     });
-}
-
-const DEFAULT_MB: usize = 4;
-
-fn read_envvar() -> Option<usize> {
-    match env::var("ESTREE_STACK_SIZE_MB") {
-        Ok(s) => {
-            match s.parse() {
-                Ok(x) => Some(x),
-                Err(_) => None
-            }
-        }
-        Err(env::VarError::NotPresent) => Some(DEFAULT_MB),
-        Err(_) => None
-    }
-}
-
-fn stack_size() -> usize {
-    let mb = match read_envvar() {
-        Some(x) => x,
-        None => {
-            println!("warning: invalid ESTREE_STACK_SIZE_MB value; defaulting to 4MB");
-            DEFAULT_MB
-        }
-    };
-    mb * 1024 * 1024
-}
-
-fn integration_tests(target: &mut Vec<TestDescAndFn>, ignore: bool) {
-    let root = Path::new(env!("CARGO_MANIFEST_DIR"));
-
-    let fixtures = {
-        let mut fixtures = root.to_path_buf();
-        fixtures.push("tests");
-        fixtures.push("esprima");
-        fixtures.push("test");
-        fixtures.push("3rdparty");
-        fixtures
-    };
-
-    let syntax = fixtures.join("syntax");
-
-    let files =
-        read_dir(syntax).unwrap()
-        .map(|dir| dir.unwrap().path())
-        .filter(|path| path.extension() == Some(OsStr::new("json")))
-        .map(|tree_path| {
-            let mut source_path_buf = fixtures.join(tree_path.file_name().unwrap());
-            source_path_buf.set_extension("js");
-            (tree_path, source_path_buf)
-        });
-
-    for (tree_path, source_path) in files {
-        let name = source_path.strip_prefix(&root).unwrap().to_str().unwrap().to_string();
-        if !ignore {
-            let mut source = String::new();
-            File::open(source_path.clone()).unwrap().read_to_string(&mut source).unwrap();
-            print!("Parsing JSON {}...", tree_path.strip_prefix(&root).unwrap().to_str().unwrap());
-            stdout().flush().unwrap();
-            let expected_ast = thread::Builder::new().stack_size(stack_size()).spawn(|| {
-                let v: Value = serde_json::de::from_reader(File::open(tree_path).unwrap()).unwrap();
-                v.into_object().unwrap().into_script().map_err(|err| {
-                    format!("failed to deserialize script: {}", err)
-                }).unwrap()
-            }).unwrap().join().unwrap();
-            println!(" done");
-            add_bench(target, name, ignore, move |mut bench| {
-                let mut result = None;
-                bench.iter(|| {
-                    result = Some(script(&source[..]))
-                });
-                match result.unwrap() {
-                    Ok(mut actual_ast) => {
-                        actual_ast.untrack();
-                        assert!(actual_ast == expected_ast, "integration test got wrong result");
-                    }
-                    Err(actual_err) => {
-                        panic!("integration test failed to parse:\n{:#?}", actual_err);
-                    }
-                }
-            });
-        } else {
-            add_bench(target, name, ignore, |_| {});
-        }
-    }
 }
 
 fn unit_tests(target: &mut Vec<TestDescAndFn>) {
@@ -231,16 +124,8 @@ fn unit_tests(target: &mut Vec<TestDescAndFn>) {
 
 fn main() {
     let args: Vec<_> = env::args().collect();
-    let bench = args.contains(&"--bench".to_string());
     let mut tests = Vec::new();
-    if !bench {
-        as_ref_test(&mut tests);
-        unit_tests(&mut tests);
-    }
-    let ignore_integration_tests = !bench && env::var_os("ESTREE_INTEGRATION_TESTS") == None;
-    if ignore_integration_tests {
-        println!("note: Run with `ESTREE_INTEGRATION_TESTS=1` to run with integration tests (much slower).");
-    }
-    integration_tests(&mut tests, ignore_integration_tests);
+    as_ref_test(&mut tests);
+    unit_tests(&mut tests);
     test_main(&args, tests);
 }
